More design documentation
Still struggling with testing anything that goes through the
Foundation/GamesByEmail framework. Hopefully, I can minimize this issue
by pushing most reasoning into Data.js.

Note that the right way to include Data.js seems to be to ask for it in
Default.htm. Not doing so now b/c Data.js is not ready for primetime.